@@ -2,7 +2,7 @@
  *
  * See: https://expressjs.com/en/4x/api.html#router, https://expressjs.com/en/guide/routing.html
  *
- * Copyright (c) 2021-2022 Julian Knight (Totally Information)
+ * Copyright (c) 2021-2023 Julian Knight (Totally Information)
  * https://it.knightnet.org.uk, https://github.com/TotallyInformation/node-red-contrib-uibuilder
  *
  * Licensed under the Apache License, Version 2.0 (the 'License');
@@ -266,7 +266,7 @@ function adminRouterV3(uib, log) {
                     log.trace(`[uibuilder:adminRouterV3:GET] Admin API. List all folders. url=${params.url}, root fldr=${uib.rootFolder}`)
 
                     // get list of all (sub)folders (follow symlinks as well)
-                    //const out = { 'root': [] }
+                    // const out = { 'root': [] }
                     const out = []
                     const root2 = uib.rootFolder.replace(/\\/g, '/')
                     fg.stream(
@@ -413,8 +413,8 @@ function adminRouterV3(uib, log) {
 
                 case 'updatepackage': {
                     log.trace(`[uibuilder:adminRouterV3:PUT:updatepackage] url=${params.url}`)
-                    console.log(`[uibuilder:adminRouterV3:PUT:updatepackage] url=${params.url}, pkg=${params.pkgName}`)
-                    
+                    // console.log(`[uibuilder:adminRouterV3:PUT:updatepackage] url=${params.url}, pkg=${params.pkgName}`)
+
                     res.statusMessage = 'PUT successful'
                     res.status(200).json({
                         newVersion: ''
