@@ -2,7 +2,7 @@
 /**
  * General utility library for Node.JS
  *
- * Copyright (c) 2019-2022 Julian Knight (Totally Information)
+ * Copyright (c) 2019-2023 Julian Knight (Totally Information)
  * https://it.knightnet.org.uk
  *
  * Licensed under the Apache License, Version 2.0 (the 'License');
