@@ -1,6 +1,6 @@
 /** Template Node-RED node runtime
  *
- * Copyright (c) 2022 Julian Knight (Totally Information)
+ * Copyright (c) 2022-2023 Julian Knight (Totally Information)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
