@@ -28,7 +28,7 @@ module.exports = {
     },
     overrides: [
         {
-            files: ['*.mod.js', '*.mjs'],
+            files: ['*.module.js', '*.mod.js', '*.mjs'],
             parserOptions: { sourceType: 'module' },
         }
     ],
