@@ -1,52 +1,8 @@
-/* eslint-disable strict */
-/* jshint browser: true, esversion: 6, asi: true */
-/* globals uibuilder */
-// @ts-nocheck
+/** The simplest use of uibuilder client library
+ * See the docs if the client doesn't start on its own.
+ */
 
-/** Minimalist code for uibuilder and Node-RED */
-'use strict'
-
-// return formatted HTML version of JSON object
-window.syntaxHighlight = function (json) {
-    json = JSON.stringify(json, undefined, 4)
-    json = json.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;')
-    json = json.replace(/("(\\u[a-zA-Z0-9]{4}|\\[^u]|[^\\"])*"(\s*:)?|\b(true|false|null)\b|-?\d+(?:\.\d*)?(?:[eE][+\-]?\d+)?)/g, function (match) {
-        var cls = 'number'
-        if ((/^"/).test(match)) {
-            if ((/:$/).test(match)) {
-                cls = 'key'
-            } else {
-                cls = 'string'
-            }
-        } else if ((/true|false/).test(match)) {
-            cls = 'boolean'
-        } else if ((/null/).test(match)) {
-            cls = 'null'
-        }
-        return '<span class="' + cls + '">' + match + '</span>'
-    })
-    return json
-} // --- End of syntaxHighlight --- //
-
-// Send a message back to Node-RED
-window.fnSendToNR = function fnSendToNR(payload) {
-    uibuilder.send({
-        'topic': 'msg-from-uibuilder-front-end',
-        'payload': payload,
-    })
-}
-
-// run this function when the document is loaded
-window.onload = function() {
-    // Start up uibuilder - see the docs for the optional parameters
-    uibuilder.start()
-
-    // Listen for incoming messages from Node-RED
-    uibuilder.onChange('msg', function(msg){
-        console.info('[indexjs:uibuilder.onChange] msg received from Node-RED server:', msg)
-
-        // dump the msg as text to the "msg" html element
-        const eMsg = document.getElementById('msg')
-        eMsg.innerHTML = window.syntaxHighlight(msg)
-    })
-}
+// Listen for incoming messages from Node-RED and action
+// uibuilder.onChange('msg', (msg) => {
+//     // do stuff with the incoming msg
+// })
