@@ -1,3 +1,4 @@
+// @ts-nocheck
 import svelte from 'rollup-plugin-svelte'
 import commonjs from '@rollup/plugin-commonjs'
 import resolve from '@rollup/plugin-node-resolve'
@@ -9,84 +10,87 @@ import css from 'rollup-plugin-css-only'
 const uibDist = 'dist'
 
 // Assume production mode if not running dev
-const production = !process.env.ROLLUP_WATCH;
-console.log(`Production mode?: ${ production }. Output Folder: ${__dirname}/${uibDist}/build/`)
+const production = !process.env.ROLLUP_WATCH
+console.log(`Production mode?: ${ production }. Output Folder: ${__dirname}/${uibDist}/build/`) // eslint-disable-line no-undef
 
+/** Define the dev server
+ * @returns {Function} writeBundle
+ */
 function serve() {
-	let server;
+    let server
 
-	function toExit() {
-		if (server) server.kill(0);
-	}
+    function toExit() { // eslint-disable-line jsdoc/require-jsdoc
+        if (server) server.kill(0)
+    }
 
-	return {
-		writeBundle() {
-			if (server) return;
-			server = require('child_process').spawn('npm', ['run', 'start', '--', '--dev'], {
-				stdio: ['ignore', 'inherit', 'inherit'],
-				shell: true
-			});
+    return {
+        writeBundle() {
+            if (server) return
+            server = require('child_process').spawn('npm', ['run', 'start', '--', '--dev'], {
+                stdio: ['ignore', 'inherit', 'inherit'],
+                shell: true
+            })
 
-			process.on('SIGTERM', toExit);
-			process.on('exit', toExit);
-		}
-	};
+            process.on('SIGTERM', toExit)
+            process.on('exit', toExit)
+        }
+    }
 }
 
 export default {
-	input: 'src/main.js',
-	output: {
-		sourcemap: true,
-		format: 'iife',
-		name: 'app',
-		file: `${uibDist}/build/bundle.js`
-	},
-	plugins: [
-		svelte({
-			compilerOptions: {
-				// enable run-time checks when not in production
-				dev: !production
-			}
-		}),
-		// we'll extract any component CSS out into
-		// a separate file - better for performance
-		css({ output: 'bundle.css' }),
+    input: 'src/main.js',
+    output: {
+        sourcemap: true,
+        format: 'iife',
+        name: 'app',
+        file: `${uibDist}/build/bundle.js`,
+    },
+    plugins: [
+        svelte({
+            compilerOptions: {
+                // enable run-time checks when not in production
+                dev: !production
+            }
+        }),
+        // we'll extract any component CSS out into
+        // a separate file - better for performance
+        css({ output: `${uibDist}/build/bundle.css` }),
 
-		// If you have external dependencies installed from
-		// npm, you'll most likely need these plugins. In
-		// some cases you'll need additional configuration -
-		// consult the documentation for details:
-		// https://github.com/rollup/plugins/tree/master/packages/commonjs
-		resolve({
-			browser: true,
-			dedupe: ['svelte']
-		}),
-		commonjs(),
+        // If you have external dependencies installed from
+        // npm, you'll most likely need these plugins. In
+        // some cases you'll need additional configuration -
+        // consult the documentation for details:
+        // https://github.com/rollup/plugins/tree/master/packages/commonjs
+        resolve({
+            browser: true,
+            dedupe: ['svelte']
+        }),
+        commonjs(),
 
-		// In dev mode, call `npm run start` once
-		// the bundle has been generated
-		!production && serve(),
+        // In dev mode, call `npm run start` once
+        // the bundle has been generated
+        !production && serve(),
 
-		// Watch the `src` directory and refresh the
-		// browser on changes when not in production
-		!production && livereload(uibDist),
+        // Watch the `src` directory and refresh the
+        // browser on changes when not in production
+        !production && livereload(uibDist),
 
-		// If we're building for production (npm run build
-		// instead of npm run dev), minify
-		production && terser({ 
-			ecma: 2015, // ES6
-			mangle: { toplevel: true },
-			compress: {
-				module: true,
-				toplevel: true,
-				unsafe_arrows: true,
-				drop_console: production,
-				drop_debugger: production,
-          	},
-			output: { comments: false } ,
-		})
-	],
-	watch: {
-		clearScreen: false
-	}
-};
+        // If we're building for production (npm run build
+        // instead of npm run dev), minify
+        production && terser({
+            ecma: 2015, // ES6
+            mangle: { toplevel: true },
+            compress: {
+                module: true,
+                toplevel: true,
+                unsafe_arrows: true,
+                drop_console: production,
+                drop_debugger: production,
+            },
+            output: { comments: false } ,
+        }),
+    ],
+    watch: {
+        clearScreen: false,
+    },
+}
